Removed floating action button. New Expense and Income now added through the nav drawer. Logout button now works in the nav drawer. Input types set for login fields.
