+++ app/src/main/java/com/hozella/budgetingessentials/DashboardFragment.java
@@ -32,27 +32,8 @@ import java.util.Date;
 
 public class DashboardFragment extends Fragment {
 
-    // Floating action button
-    private FloatingActionButton fab_main;
-    private FloatingActionButton fab_income;
-    private FloatingActionButton fab_expense;
-
-    // Floating action button text
-    private TextView fab_income_txt;
-    private TextView fab_expense_txt;
-
     private TextView expense_total;
     private TextView income_total;
-
-    public static boolean isOpen = false;
-
-    // Floating action button animation
-    private Animation fadeOpen, fadeClose;
-
-    // Firebase
-    private FirebaseAuth mAuth;
-    private DatabaseReference mIncomeDatabase;
-    private DatabaseReference mExpenseDatabase;
 
     public DashboardFragment() {
 
@@ -64,12 +45,7 @@ public class DashboardFragment extends Fragment {
         // Inflate the layout for this fragment
         View myView = inflater.inflate(R.layout.fragment_dashboard, container, false);
 
-        mAuth = FirebaseAuth.getInstance();
-        FirebaseUser mUser = mAuth.getCurrentUser();
-        String uid = mUser.getUid();
 
-        mIncomeDatabase = FirebaseDatabase.getInstance().getReference().child("IncomeData").child(uid);
-        mExpenseDatabase = FirebaseDatabase.getInstance().getReference().child("ExpenseData").child(uid);
 
         // Update Income and Expense total amounts
         expense_total = myView.findViewById(R.id.expense_dashboard_total);
@@ -77,224 +53,12 @@ public class DashboardFragment extends Fragment {
         expense_total.setText(String.valueOf(HomeActivity.expenseTotalSum));
         income_total.setText(String.valueOf(HomeActivity.incomeTotalSum));
 
-        // Connect floating buttons and texts to layout
-        fab_main = myView.findViewById(R.id.fb_main_btn);
-        fab_income = myView.findViewById(R.id.income_ft_btn);
-        fab_expense = myView.findViewById(R.id.expense_ft_btn);
-        fab_income_txt = myView.findViewById(R.id.income_ft_text);
-        fab_expense_txt = myView.findViewById(R.id.expense_ft_text);
-
-        // Animations
-        fadeOpen = AnimationUtils.loadAnimation(getActivity(), R.anim.fade_open);
-        fadeClose = AnimationUtils.loadAnimation(getActivity(), R.anim.fade_close);
-
-        fab_main.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                addData();
-                ftAnimation();
-            }
-        });
-
         return myView;
     }
 
-    // Floating Button Animation
-    private void ftAnimation(){
-        if (isOpen){
-            fab_income.startAnimation(fadeClose);
-            fab_expense.startAnimation(fadeClose);
-            fab_income.setClickable(false);
-            fab_expense.setClickable(false);
-
-            fab_income_txt.startAnimation(fadeClose);
-            fab_expense_txt.startAnimation(fadeClose);
-            fab_income_txt.setClickable(false);
-            fab_expense_txt.setClickable(false);
-            isOpen = false;
-
-        }else {
-            fab_income.startAnimation(fadeOpen);
-            fab_expense.startAnimation(fadeOpen);
-            fab_income.setClickable(true);
-            fab_expense.setClickable(true);
-
-            fab_income_txt.startAnimation(fadeOpen);
-            fab_expense_txt.startAnimation(fadeOpen);
-            fab_income_txt.setClickable(true);
-            fab_expense_txt.setClickable(true);
-            isOpen = true;
-        }
-    }
-
-    private void addData(){
-
-        // Fab Button Income
-        fab_income.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                incomeDataInsert();
-            }
-        });
-
-        // Fab Button Expense
-        fab_expense.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                expenseDataInsert();
-            }
-        });
-
-    }
-
-    public void incomeDataInsert(){
-        AlertDialog.Builder myDialog = new AlertDialog.Builder(getActivity());
-
-        LayoutInflater inflater = LayoutInflater.from(getActivity());
-
-        View myView = inflater.inflate(R.layout.layout_insert_data, null);
-        myDialog.setView(myView);
-        AlertDialog dialog = myDialog.create();
-        dialog.setCancelable(false);
-
-        EditText edtAmount = myView.findViewById(R.id.amount_edt);
-        EditText edtType = myView.findViewById(R.id.type_edt);
-        EditText edtNote = myView.findViewById(R.id.note_edt);
-
-        Button btnSave = myView.findViewById(R.id.btnSave);
-        Button btnCancel = myView.findViewById(R.id.btnCancel);
-
-        btnSave.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String type = edtType.getText().toString().trim();
-                String amount = edtAmount.getText().toString().trim();
-                String note = edtNote.getText().toString().trim();
-
-                if(TextUtils.isEmpty(type)){
-                    edtType.setError("Required Field...");
-                    return;
-                }
-
-                if(TextUtils.isEmpty(amount)){
-                    edtAmount.setError("Required Field...");
-                    return;
-                } else if(!TextUtils.isDigitsOnly(amount)){
-                    edtAmount.setError("Amount must be a number!");
-                    return;
-                    }
-
-
-
-                String id = mIncomeDatabase.push().getKey();
-                String mDate = DateFormat.getDateInstance().format(new Date());
-                Data data = new Data(amount, type, note, id, mDate);
-                mIncomeDatabase.child(id).setValue(data).addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        Toast.makeText(getActivity(), "Data Upload Successful", Toast.LENGTH_SHORT).show();
-                    }
-                })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Toast.makeText(getActivity(), "Data Upload Failed", Toast.LENGTH_SHORT).show();
-                                System.out.println("DATA UPLOAD FIREBASE FAILED: " + e);
-                            }
-                        });;
-
-
-                dialog.dismiss();
-                ftAnimation();
-            }
-        });
-
-        btnCancel.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                dialog.dismiss();
-                ftAnimation();
-            }
-        });
-
-        dialog.show();
-    }
-
-    public void expenseDataInsert(){
-        AlertDialog.Builder myDialog = new AlertDialog.Builder(getActivity());
-
-        LayoutInflater inflater = LayoutInflater.from(getActivity());
-
-        View myView = inflater.inflate(R.layout.layout_insert_data, null);
-        myDialog.setView(myView);
-        AlertDialog dialog = myDialog.create();
-        dialog.setCancelable(false);
-
-        EditText edtAmount = myView.findViewById(R.id.amount_edt);
-        EditText edtType = myView.findViewById(R.id.type_edt);
-        EditText edtNote = myView.findViewById(R.id.note_edt);
-
-        Button btnSave = myView.findViewById(R.id.btnSave);
-        Button btnCancel = myView.findViewById(R.id.btnCancel);
-
-        btnSave.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String type = edtType.getText().toString().trim();
-                String amount = edtAmount.getText().toString().trim();
-                String note = edtNote.getText().toString().trim();
-
-                if(TextUtils.isEmpty(type)){
-                    edtType.setError("Required Field...");
-                    return;
-                }
-
-                if(TextUtils.isEmpty(amount)){
-                    edtAmount.setError("Required Field...");
-                    return;
-                } else if(!TextUtils.isDigitsOnly(amount)){
-                    edtAmount.setError("Amount must be a number!");
-                    return;
-                    }
 
 
 
 
-
-
-
-                String id = mExpenseDatabase.push().getKey();
-                String mDate = DateFormat.getDateInstance().format(new Date());
-                Data data = new Data(amount, type, note, id, mDate);
-                mExpenseDatabase.child(id).setValue(data).addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        Toast.makeText(getActivity(), "Data Upload Successful", Toast.LENGTH_SHORT).show();
-                    }
-                })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Toast.makeText(getActivity(), "Data Upload Failed", Toast.LENGTH_SHORT).show();
-                                System.out.println("DATA UPLOAD FIREBASE FAILED: " + e);
-                            }
-                        });;
-
-
-                dialog.dismiss();
-                ftAnimation();
-            }
-        });
-
-        btnCancel.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                dialog.dismiss();
-                ftAnimation();
-            }
-        });
-
-        dialog.show();
-    }
 
 }
