--- app/src/main/res/anim/fade_close.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:fillAfter="true">
-
-    <scale android:duration="500"
-        android:fromXScale="1.0"
-        android:fromYScale="1.0"
-        android:toYScale="0.0"
-        android:toXScale="0.0"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:interpolator="@android:anim/linear_interpolator">
-
-    </scale>
-
-    <alpha android:duration="500"
-        android:fromAlpha="1.0"
-        android:toAlpha="0.0"
-        android:interpolator="@android:anim/linear_interpolator">
-
-    </alpha>
-
-</set>
