--- app/src/main/res/anim/fade_open.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:fillAfter="true">
-
-    <scale android:duration="500"
-        android:fromXScale="0.0"
-        android:fromYScale="0.0"
-        android:toYScale="1.0"
-        android:toXScale="1.0"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:interpolator="@android:anim/linear_interpolator">
-
-    </scale>
-
-    <alpha android:duration="500"
-        android:fromAlpha="0.0"
-        android:toAlpha="1.0"
-        android:interpolator="@android:anim/accelerate_interpolator">
-
-    </alpha>
-
-
-</set>
