+++ app/src/main/res/layout/activity_main.xml
@@ -42,6 +42,7 @@
                 android:layout_margin="10dp"
                 android:background="@drawable/edittext_background"
                 android:hint="email@website.com"
+                android:inputType="textEmailAddress"
                 android:minHeight="48dp"
                 android:padding="15dp"
                 android:textColor="@color/black">
@@ -55,6 +56,7 @@
                 android:layout_margin="10dp"
                 android:background="@drawable/edittext_background"
                 android:hint="Password"
+                android:inputType="textPassword"
                 android:minHeight="48dp"
                 android:padding="15dp"
                 android:textColor="@color/black">
